UBUNTU: SAUCE: apparmor4.0.0 [83/84]: apparmor: fix prompt failing during large down loads
Firefox was failing large downloads with a denied message, after a
download was started for two interelated reasons. The first was that
the full set of allowed permissions where not being propogated back
to the object fctx cache, which could result in secondary permission
requests outside of the regular open and create hooks.

Second the atomic_context information was not being correctly propgated
so that some of those secondary checks would not result in a prompt
but a default deny.

Signed-off-by: John Johansen <john.johansen@canonical.com>