UBUNTU: [Packaging] rules: Fold module signature check into stamp-install-% target
We currently have two phony targets checks-% and module-signature-check-%.
The first depends on the second and the second does nothing other than
checking the module signatures. With a dependency chain like:

binary-% <- checks-% <- module-signature-check-% <- stamps/stamp-install-%

This is a left-over from the days when we had multiple checks. We don't
need nor use these phony targets, they only add unnecessary complexity.
Just run the check in the install step and get rid of the phony targets.
This simplifies the chain to:

binary-% <- stamps/stamp-install-%

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>