UBUNTU: SAUCE: apparmor4.0.0 [40/53]: apparmor: Document that label must be last member in struct aa_profile
BugLink: http://bugs.launchpad.net/bugs/2028253

The label struct is variable length. While its use in struct aa_profile
is fixed length at 2 entries the variable length member needs to be
the last member in the structure.

The code already does this but the comment has it in the wrong location.
Also add a comment to ensure it stays at the end of the structure.

While we are at it, update the documentation for other profile members
as well.

Signed-off-by: John Johansen <john.johansen@canonical.com>