UBUNTU: SAUCE: gpio: aggregator: reorder functions to prepare for configfs introduction
BugLink: https://bugs.launchpad.net/bugs/2103496

Reorder functions in drivers/gpio/gpio-aggregator.c to prepare for the
configfs-based interface additions in subsequent commits. Arrange the
code so that the configfs implementations will appear above the existing
sysfs-specific code, since the latter will partly depend on the configfs
interface implementations when it starts to expose the settings to
configfs.

The order in drivers/gpio/gpio-aggregator.c will be as follows:

* Basic gpio_aggregator/gpio_aggregator_line representations
* Common utility functions
* GPIO Forwarder implementations
* Configfs interface implementations
* Sysfs interface implementations
* Platform device implementations
* Module init/exit implementations

This separate commit ensures a clean diff for the subsequent commits.

No functional change.

Reviewed-by: Geert Uytterhoeven <geert+renesas@glider.be>
[koichiroden: cherry-pick [PATCH v6 2/9] from:
https://lore.kernel.org/all/20250315164123.1855142-1-koichiro.den@canonical.com/]
Signed-off-by: Koichiro Den <koichiro.den@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>